Generify the input handler logic making use of resource handlers and remove the concept of long input handlers, as chemicals can no longer extract at rates higher than max int

+++ src/main/java/mekanism/common/recipe/lookup/IRecipeLookupHandler.java
@@ -100,7 +100,7 @@ public interface IRecipeLookupHandler<RECIPE extends MekanismRecipe<?>> extends 
          *
          * @return Constant amount of secondary input that had been used before saving.
          */
-        default long getSavedUsedSoFar(int cacheIndex) {
+        default int getSavedUsedSoFar(int cacheIndex) {
             return 0;
         }
     }
